Marconi.Cardano.ChainIndex.Api.JsonRpc.Endpoint.TargetAddresses

Documentation

type RpcTargetAddressesMethod = JsonRpc "getTargetAddresses" UnusedRequestParams String [Text]

getTargetAddressesQueryHandler

Arguments

:: UnusedRequestParams
Will be an empty string, empty object, or null, as we are ignoring this param, and returning everything

-> ReaderHandler HttpServerConfig (Either (JsonRpcErr String) [Text])

Return the list of TargetAddresses in Bech32 representation.