Marconi.Cardano.ChainIndex.CLI

Documentation

data StartingPoint

Represents a specified point from which to start indexing

Constructors

StartFromGenesis

StartFromLastSyncPoint

StartFrom ChainPoint

commonStartFromParser :: Parser StartingPoint

Allow the user to set a starting point for indexing.

If the user picks a specific point with the start-from command, they need to provide both a BlockHeaderHash (encoded in RawBytesHex) and a SlotNo (a natural number).

If the chooses either --start-from-genesis or --start-from-last-sync-points, no args are required.

Not using any of these commands results in the same behaviour as --start-from-last-sync-points

commonNetworkIdParser :: Parser NetworkId

pMainnetParser :: Parser NetworkId

pTestnetMagicParser :: Parser NetworkMagic

multiAddressesParser :: Mod OptionFields [Address ShelleyAddr] -> Parser TargetAddresses

parses CLI params to valid NonEmpty list of Shelley addresses We error out if there are any invalid addresses

data CommonOptions

This executable is meant to exercise a set of indexers (for now datumhash -> datum) against the mainnet (meant to be used for testing).

In case you want to access the results of the datumhash indexer you need to query the resulting database: $ sqlite3 datums.sqlite > select slotNo, datumHash, datum from kv_datumhsh_datum where slotNo = 39920450; 39920450|679a55b523ff8d61942b2583b76e5d49498468164802ef1ebe513c685d6fb5c2|X(002f9787436835852ea78d3c45fc3d436b324184

Constructors

CommonOptions

Fields

optionsSocketPath :: !String

POSIX socket file to communicate with cardano node

optionsNetworkId :: !NetworkId

cardano network id

optionsChainPoint :: !StartingPoint

The starting point of the indexers

optionsRetryConfig :: !RetryConfig

data Options

Constructors

Options

Fields

commonOptions :: !CommonOptions

optionsDbPath :: !FilePath

Directory path containing the SQLite database files

optionsEnableUtxoTxOutRef :: !Bool

enable storing txout refScript,

optionsDisableUtxo :: !Bool

disable Utxo indexer

optionsDisableAddressDatum :: !Bool

disable AddressDatum indexer

optionsDisableScript :: !Bool

disable Script indexer

optionsDisableEpochState :: !Bool

disable EpochState indexer

optionsDisableMintBurn :: !Bool

disable MintBurn indexer

optionsRpcPort :: !Int

port the RPC server should listen on

optionsTargetAddresses :: !(Maybe TargetAddresses)

white-space separated list of Bech32 Cardano Shelley addresses

optionsTargetAssets :: !(Maybe (NonEmpty (PolicyId, Maybe AssetName)))

white-space separated list of target asset policy id and optionally asset name, separated by ..

optionsNodeConfigPath :: !(Maybe FilePath)

Path to the node config

parseOptions :: IO Options

parseSnapshotOptions :: IO SnapshotOptions

getVersion :: String

programParser :: ParserInfo Options

snapshotProgramParser :: ParserInfo SnapshotOptions

commonOptionsParser :: Parser CommonOptions

optionsParser :: Parser Options

Database paths

utxoDbPath :: Options -> Maybe FilePath

addressDatumDbPath :: Options -> Maybe FilePath

scriptTxDbPath :: Options -> Maybe FilePath

epochStateDbPath :: Options -> Maybe FilePath

mintBurnDbPath :: Options -> Maybe FilePath

Common CLI parsers for other derived programs.

commonSocketPathParser :: Parser String

commonDbDirParser :: Parser String

Root directory for the SQLite storage of all the indexers

snapshotDirParser :: Parser String

commonVersionOptionParser :: Parser (a -> a)

marconiDescr :: String -> InfoMod a

commonPortParser :: Parser Int

commonMaybeTargetAddressParser :: Parser (Maybe TargetAddresses)

Parse the addresses to index. Addresses should be given in Bech32 format Several addresses can be given in a single string, if they are separated by a space

commonMaybeTargetAssetParser :: Parser (Maybe (NonEmpty (PolicyId, Maybe AssetName)))

Parse target assets, both the PolicyId and the AssetName are expected to be in their RawBytesHex representation, they must be separated by a comma. The asset name can be omited, if it is the case, any asset with the expected policy ID will be matched. Several assets can be given in a single string if you separate them with a space.

parseAsset :: Text -> ReadM (PolicyId, Maybe AssetName)

Asset parser, see commonMaybeTargetAssetParser for more info.

commonNodeConfigPathParser :: Parser FilePath

commonRetryConfigParser :: Parser RetryConfig

Allow the user to specify the retry config when the connection to the node is lost.

mkUtxoIndexerConfig :: Options -> UtxoIndexerConfig

Extract UtxoIndexerConfig from CLI Options

data SnapshotOptions

CL options for the marconi-chain-snapshot executable.

Constructors

SnapshotOptions

Fields

snapshotOptionsSocketPath :: !String

POSIX socket file to communicate with cardano node

snapshotOptionsNetworkId :: !NetworkId

Cardano network id

snapshotOptionsRetryConfig :: !RetryConfig

Time to wait until retrying socket connection

snapshotOptionsSnapshotDir :: !FilePath

Directory path containing the files for each snapshot

snapshotOptionsNodeConfigPath :: !(Maybe FilePath)

Path to the node config

snapshotOptionsBlockRanges :: ![BlockRange]

A list of block ranges to snapshot

snapshotOptionsParser :: Parser SnapshotOptions

snapshotBlockRangesParser :: Parser [BlockRange]

readBlockRange :: String -> ReadM BlockRange