Marconi.Cardano.ChainIndex.Indexers

Documentation

type Coordinator = WithTrace IO Coordinator TipAndBlock

type SyncStatsCoordinator = WithSyncStats (WithTrace IO Coordinator) TipAndBlock

type ChainTipIndexer = ChainTipIndexer IO

type MintTokenEventIndexer = StandardIndexer IO SQLiteIndexer MintTokenBlockEvents

type BlockInfoIndexer = StandardIndexer IO SQLiteIndexer BlockInfo

type UtxoIndexer = UtxoQueryIndexer IO

type CurrentSyncPointIndexer = WithTrace IO CurrentSyncPointQueryIndexer TipAndBlock

type EpochNonceIndexer = WithTrace IO SQLiteIndexer EpochNonce

type EpochSDDIndexer = WithTrace IO SQLiteIndexer (NonEmpty EpochSDD)

data EpochEvent

Constructors

EpochEvent

Fields

epochNo :: EpochNo

epochSDD :: Maybe (NonEmpty EpochSDD)

epochNonce :: Maybe EpochNonce

data MarconiCardanoQueryables

Container for all the queryable indexers.

Constructors

MarconiCardanoQueryables

Fields

_queryableEpochNonce :: !(MVar EpochNonceIndexer)

_queryableEpochSDD :: !(MVar EpochSDDIndexer)

_queryableMintToken :: !(MintTokenEventIndexerQuery MintTokenBlockEvents)

_queryableUtxo :: !UtxoIndexer

_queryableCurrentSyncPoint :: !CurrentSyncPointIndexer

queryableUtxo :: Lens' MarconiCardanoQueryables UtxoIndexer

queryableMintToken :: Lens' MarconiCardanoQueryables (MintTokenEventIndexerQuery MintTokenBlockEvents)

queryableEpochSDD :: Lens' MarconiCardanoQueryables (MVar EpochSDDIndexer)

queryableEpochNonce :: Lens' MarconiCardanoQueryables (MVar EpochNonceIndexer)

queryableCurrentSyncPoint :: Lens' MarconiCardanoQueryables CurrentSyncPointIndexer

buildIndexers :: SecurityParam -> CatchupConfig -> UtxoIndexerConfig -> MintTokenEventConfig -> ExtLedgerStateWorkerConfig EpochEvent (WithDistance BlockEvent) -> Trace IO Text -> MarconiTrace IO -> FilePath -> ExceptT IndexerError IO (ChainPoint, MarconiCardanoQueryables, SyncStatsCoordinator)

Build all the indexers and expose a single coordinator to operate them

buildBlockEventCoordinator :: MonadIO m => Trace IO (IndexerEvent ChainPoint) -> [Worker (WithDistance BlockEvent) ChainPoint] -> m (Worker TipAndBlock ChainPoint)

Build and start a coordinator of a bunch of workers that takes an AnyTxBody as an input

buildTxBodyCoordinator :: (MonadIO m, Ord (Point event), Show (Point event)) => Trace IO Text -> (WithDistance input -> IO (Maybe event)) -> [Worker event (Point event)] -> m (Worker (WithDistance input) (Point event))

Build and start a coordinator of a bunch of workers that takes an AnyTxBody as an input

epochNonceBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => SecurityParam -> CatchupConfig -> Trace m Text -> FilePath -> n (WorkerIndexer m EpochEvent EpochNonce (WithTrace m SQLiteIndexer))

Configure and start the EpochNonce indexer

epochSDDBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => SecurityParam -> CatchupConfig -> Trace m Text -> FilePath -> n (WorkerIndexer m EpochEvent (NonEmpty EpochSDD) (WithTrace m SQLiteIndexer))

Configure and start the EpochSDD indexer

snapshotBlockEventBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => SecurityParam -> CatchupConfig -> Trace m Text -> FilePath -> BlockRange -> FilePath -> n (WorkerIndexer m (ExtLedgerStateEvent, WithDistance BlockEvent) SnapshotBlockEvent (WithTrace m (FileIndexer SnapshotMetadata)))

Configure and start the SnapshotBlockEvent indexer

extractSnapshotBlockEvent :: Applicative f => (ExtLedgerStateEvent, WithDistance BlockEvent) -> f (Maybe SnapshotBlockEvent)

buildIndexersForSnapshot :: SecurityParam -> CatchupConfig -> ExtLedgerStateWorkerConfig (ExtLedgerStateEvent, WithDistance BlockEvent) (WithDistance BlockEvent) -> Trace IO Text -> MarconiTrace IO -> FilePath -> [BlockRange] -> FilePath -> ExceptT IndexerError IO SyncStatsCoordinator

Builds the coordinators for each sub-chain serializers.

snapshotExtLedgerStateEventBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => SecurityParam -> CatchupConfig -> Trace m Text -> FilePath -> BlockRange -> FilePath -> n (WorkerIndexer m (ExtLedgerStateEvent, WithDistance BlockEvent) ExtLedgerStateEvent (WithTrace m (FileIndexer EpochMetadata)))

snapshotExtLedgerStateEventWorker :: forall input m n. (MonadIO m, MonadError IndexerError m, MonadIO n) => StandardWorkerConfig n input ExtLedgerStateEvent -> SnapshotWorkerConfig input -> FilePath -> m (WorkerIndexer n input ExtLedgerStateEvent (WithTrace n (FileIndexer EpochMetadata)))

justBeforeBlockRangePreprocessor :: Monad m => (a -> BlockNo) -> BlockRange -> Preprocessor m ChainPoint a a